+++ packages/storybook-vue/stories/components/telekom-brand-header-navigation/SlotMainNavWithFlyout.vue
@@ -1,0 +1,135 @@
+<template>
+  <scale-telekom-nav-list variant="main-nav" slot="main-nav" aria-label="Main navigation links">
+    <scale-telekom-nav-item active aria-label="Topic One">
+      <button>Topic One</button>
+      <scale-telekom-nav-flyout hover>
+        <scale-telekom-mega-menu>
+          <scale-telekom-mega-menu-column>
+            <scale-icon-home-home slot="icon"></scale-icon-home-home>
+            <a href="https://example.com" slot="heading">Second Level</a>
+            <ul>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+            </ul>
+          </scale-telekom-mega-menu-column>
+          <scale-telekom-mega-menu-column>
+            <scale-icon-home-home slot="icon"></scale-icon-home-home>
+            <a href="https://example.com" slot="heading">Second Level</a>
+            <ul>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+            </ul>
+          </scale-telekom-mega-menu-column>
+          <scale-telekom-mega-menu-column>
+            <scale-icon-home-home slot="icon"></scale-icon-home-home>
+            <a href="https://example.com" slot="heading">Second Level</a>
+            <ul>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+            </ul>
+          </scale-telekom-mega-menu-column>
+          <scale-telekom-mega-menu-column>
+            <scale-icon-home-home slot="icon"></scale-icon-home-home>
+            <a href="https://example.com" slot="heading">Second Level</a>
+            <ul>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+            </ul>
+          </scale-telekom-mega-menu-column>
+          <scale-telekom-mega-menu-column>
+            <scale-icon-home-home slot="icon"></scale-icon-home-home>
+            <a href="https://example.com" slot="heading">Second Level</a>
+            <ul>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+            </ul>
+          </scale-telekom-mega-menu-column>          
+        </scale-telekom-mega-menu>
+      </scale-telekom-nav-flyout>
+    </scale-telekom-nav-item>
+    <scale-telekom-nav-item aria-label="Topic Two">
+      <button>Topic Two</button>
+      <scale-telekom-nav-flyout hover>
+        <scale-telekom-mega-menu>
+          <scale-telekom-mega-menu-column>
+            <scale-icon-home-home slot="icon"></scale-icon-home-home>
+            <a href="https://example.com" slot="heading">Second Level</a>
+            <ul>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+            </ul>
+          </scale-telekom-mega-menu-column>
+          <scale-telekom-mega-menu-column>
+            <scale-icon-home-home slot="icon"></scale-icon-home-home>
+            <a href="https://example.com" slot="heading">Second Level</a>
+            <ul>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+            </ul>
+          </scale-telekom-mega-menu-column>   
+          <scale-telekom-mega-menu-column>
+            <scale-icon-home-home slot="icon"></scale-icon-home-home>
+            <a href="https://example.com" slot="heading">Second Level</a>
+            <ul>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+            </ul>
+          </scale-telekom-mega-menu-column>
+          <scale-telekom-mega-menu-column>
+            <scale-icon-home-home slot="icon"></scale-icon-home-home>
+            <a href="https://example.com" slot="heading">Second Level</a>
+            <ul>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+              <li><a href="https://example.com">Third Level</a></li>
+            </ul>
+          </scale-telekom-mega-menu-column>          
+        </scale-telekom-mega-menu>
+      </scale-telekom-nav-flyout>
+    </scale-telekom-nav-item>
+    <scale-telekom-nav-item aria-label="Topic Three">
+      <a href="#">Topic Three</a>
+    </scale-telekom-nav-item>
+    <scale-telekom-nav-item aria-label="Topic Four">
+      <a href="#">Topic Four</a>
+    </scale-telekom-nav-item>
+    <scale-telekom-nav-item aria-label="Topic Five">
+      <a href="#">Topic Five</a>
+    </scale-telekom-nav-item>
+  </scale-telekom-nav-list>
+</template>